+++ pretab/embedding/language.py
@@ -2,7 +2,7 @@ import numpy as np
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.utils.validation import check_is_fitted
 
-from ...exceptions import OptionalDependencyError, PretabConfigError, PretabDataError
+from ..exceptions import OptionalDependencyError, PretabConfigError, PretabDataError
 
 
 class LanguageEmbeddingTransformer(TransformerMixin, BaseEstimator):